Created a file for configuring crontab.
The setup_cronjob.py file contains a Python script which creates a
cronjob according to the project specifications. The python-crontab
package is used to achieve this task. When the script is re-excuted
any existing cronjob for this project will be removed and the new
cronjob will replace it.

new file:   crontab-automation/setup_cronjob.py